+++ dynts/lib/__init__.py
@@ -1,7 +1,7 @@
 try:
     from cts import *
     hasextensions = True
-except:
+except ImportError, e:
     hasextensions = False
     from .fallback import *
 else:
@@ -9,6 +9,7 @@ else:
     
     
 def makeskiplist(data = None, expected_size = 100, use_fallback = False):
+    '''Create a new skiplist'''
     if hasattr(data,'__len__'):
         expected_size = max(expected_size,len(data))
     sl = fallback.skiplist if use_fallback else skiplist