+++ dynts/lib/fallback/simplefunc.py
@@ -1,0 +1,17 @@
+
+from .common import *
+
+def tsminmax(v):
+    mv = NaN
+    xv = NaN
+    for v in x:
+        if x == x:
+            if mv == mv:
+                mv = min(mv,x)
+            else:
+                mv = x
+            if xv == xv:
+                xv = max(xv,x)
+            else:
+                xv = x
+    return (mv,xv)