+++ dynts/tests/regression/tszoo/tests.py
@@ -1,0 +1,11 @@
+'''Timeseries Zoo R package implementation'''
+from regression import tsinterface, tsop
+
+
+class TestNumpyTS(tsinterface.TestTS):
+    backend = 'zoo'
+
+
+class TestOpNumpyTS(tsop.TestOperators):
+    backend = 'zoo'
+