+++ lib/src/skiplist.pyx
@@ -6,26 +6,7 @@
 
 # Cython version: Wes McKinney
 
-cdef extern from "numpy/arrayobject.h":
-
-    void import_array()
-
-cdef extern from "math.h":
-    double log(double x)
-
-# MSVC does not have log2!
-
-cdef double Log2(double x):
-    return log(x) / log(2.)
-
-cimport numpy as np
-from numpy cimport *
-import numpy as np
-
 from random import random
-
-# initialize numpy
-import_array()
 
 cdef class Node: