+++ CHANGELOG.rst
@@ -1,17 +1,19 @@
 
-Version 0.3.4 - Development
+Version 0.4.0 - Development
 ================================
+* Added a ``profile`` suite.
+* Added fallback tests and split ``zoo`` and ``numpy`` timeseries backend tests.
 * Added sphinx_ extension in mod:`dynts.web.dyntsx` for displaying the list of timeseries functions available.
 * Created the ``lib`` module where cython_ extensions will be placed. Cython extensions will be compiled only
   if cython is available, otherwise a fall-back pure python library will be used.
 * Redesign of test suite. You can now run tests for given tags::
 
-	python runtests dsl interface ...
+	python runtests dsl tsnumpy ...
 	
   Each tag is defined by a directory in the :mod:`dynts.tests.regression` module.
 * Added simple statistics pivot table.
 * Bug fixes.
-* **76 Tests**, **73% test coverage**
+* **80 Tests**, **73% test coverage**
 
 Version 0.3.3 - 2010 November 16
 ===================================